Migrate from scoped_model to provider
This is the minimal possible change. It is meant to be readable as a standalone Pull Request.

This does not touch comments. Additionally, it doesn’t use the niceties of provider, like automatic disposal of the model. Lastly, it doesn’t even rename the project from `scoped_model_counter`. All that will come in a follow-up PR.