[provider_shopper] migrate to null safety (#736)
This change migrates the Provider Shopper code to null safety. Since null safety isn't the main topic of the sample, the changes are made in a way that focuses attention to state management. (E.g. some parts might be made more null safe, but we didn't go there because it would take attention away from the main message.)

Co-authored-by: Filip Hracek <filip.hracek@gmail.com>