Fix go_router navigation
Mixing declarative (`.go()`) and imperative (`.pop()`) navigation no longer works.

This CL changes all navigation to the declarative `go()` wherever possible. The only imperative navigation still standing is the `/settings` route, which needs to be accessed (pushed) from both the main screen and the play session screen, and therefore there is no single URL to go back to. So we’re keeping `push()` / `pop()` there.

I’m sure it’s possible to implement even the `/settings` route with declarative navigation, but it seems way beyond the scope of this sample.

Fixes #1514.