navigation_and_routing: Handle initial route before login (Issue #947 )
This commit introduces a new functionality in the navigation_and_routing module. Now, if a user tries to access a specific route before logging in, the application will remember this initial route. After the user logs in, they will be redirected back to this initial route instead of a default route.