m3_demo - Enables edge to edge support. (#2441)
Edge to edge is coming by default to Flutter 3.26+ (or thereabouts) when
the targetSdk is raised to Android 15(API 35). In the meantime, it can
be done manually by adjusting the theme and setting
`WindowCompat.setDecorFitsSystemWindows`.

The fix is purely on the Android side. iOS is unaffected.

Sourced from
https://developer.android.com/develop/ui/views/layout/edge-to-edge-manually

Fixes #1720 .

## Pre-launch Checklist

- [X] I read the [Flutter Style Guide] _recently_, and have followed its
advice.
- [X] I signed the [CLA].
- [X] I read the [Contributors Guide].
- [X] I updated/added relevant documentation (doc comments with `///`).
- [x] All existing and new tests are passing.

If you need help, consider asking for advice on the #hackers-devrel
channel on [Discord].

<!-- Links -->
[Flutter Style Guide]:
https://github.com/flutter/flutter/blob/master/docs/contributing/Style-guide-for-Flutter-repo.md
[CLA]: https://cla.developers.google.com/
[Discord]:
https://github.com/flutter/flutter/blob/master/docs/contributing/Chat.md
[Contributors Guide]:
https://github.com/flutter/samples/blob/main/CONTRIBUTING.md